requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
beautifulsoup4==4.13.4
blinker==1.9.0
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
dataclasses-json==0.6.7
distro==1.9.0
duckduckgo_search==8.0.2
faiss-cpu==1.11.0
frozenlist==1.6.0
gitdb==4.0.12
GitPython==3.1.44
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
Jinja2==3.1.6
jiter==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
langchain==0.3.25
langchain-community==0.3.24
langchain-core==0.3.60
langchain-openai==0.3.17
langchain-tavily==0.1.6
langchain-text-splitters==0.3.8
langsmith==0.3.42
lxml==5.4.0
MarkupSafe==3.0.2
marshmallow==3.26.1
multidict==6.4.4
mypy==1.15.0
mypy_extensions==1.1.0
narwhals==1.40.0
numpy==2.2.6
openai==1.81.0
orjson==3.10.18
packaging==24.2
pandas==2.2.3
pillow==11.2.1
primp==0.15.0
propcache==0.3.1
protobuf==6.31.0
pyarrow==20.0.0
pydantic==2.11.4
pydantic-settings==2.9.1
pydantic_core==2.33.2
pydeck==0.9.1
pypdf==5.5.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rpds-py==0.25.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.41
streamlit==1.45.1
tenacity==9.1.2
tiktoken==0.9.0
toml==0.10.2
tornado==6.5
tqdm==4.67.1
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
yarl==1.20.0
zstandard==0.23.0
